// Source/SelfService/Web/apis/integrations/generated/models/AccountAccess.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aigonix.Bridge.M3
 * Bridge API - made for Dolittle Studio
 *
 * The version of the OpenAPI document: 1.0.0.0
 * Contact: dolittle@dolittle.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const AccountAccess = {
    Read: 'Read',
    ReadWrite: 'ReadWrite',
    Debug: 'Debug'
} as const;
export type AccountAccess = typeof AccountAccess[keyof typeof AccountAccess];


export function AccountAccessFromJSON(json: any): AccountAccess {
    return AccountAccessFromJSONTyped(json, false);
}

export function AccountAccessFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountAccess {
    return json as AccountAccess;
}

export function AccountAccessToJSON(value?: AccountAccess | null): any {
    return value as any;
}